Add using fixed IP to command 'workspace-desktop create'.
To create a new workspace desktop using a pre-defined, fixed IP
address (instead of automatically assigning a random IP address),
a new option "--fixed-ip" was added to specify an IP address.

To make this work, the corresponding subnet ID is also required.
To be able to specify the subnet name instead of the subnet ID,
a new option "--vpc-name" was added which is needed to determine
the subnet ID from the subnet name.