Fix command 'serverbackuppolicy remove'.
If a server was deleted using the OTC console, it will not be removed
from a backup policy that it was associated with (although the server
that was deleted won't be shown anymore in the OTC console).

However, adding or removing servers to that backup policy using the
command "serverbackuppolicy add/remove" will always fail now, because
determining the UUID for the removed server will now always fail.

This is fixed by not failing (but only warning) if determining the
UUID for a server to be removed from a backup policy fails.