Fix 'workspace-desktop list|show all' for projects without workspace.
If a project has not applied for workspace yet, attempting to list all
workspace desktops (using commands "otc workspace-desktop list" or
"otc workspace-desktop show all") resulted in an error message, as the
returned JSON containing the error message was parsed for desktop list
output, which failed. This is now handled correctly.